backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.11.3
openai==1.75.0
langdetect==1.0.9
python-multipart==0.0.9
